requirements.txt
# LangChain minimal stack (solo lo que usamos en RAG base)
langchain-core==0.3.83
langchain-community==0.3.21
langchain-text-splitters==0.3.9
langchain-huggingface==0.3.1
langchain-chroma==0.2.6

chromadb>=0.5.0
pypdf>=4.2.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
reportlab>=4.2.0

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.35.0
requests>=2.32.0
